requirements.txt
# Core Dependencies
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.3.0
seaborn>=0.11.0
scipy>=1.7.0

# Main Package
gmx_MMPBSA>=1.6.0

# YAML Support
PyYAML>=6.0

# Additional Dependencies
tqdm
parmed>=4.2.2
mpi4py<=3.1.5

# Conda-only Dependencies (설치 명령어)
# conda install -c conda-forge ambertools=23 -y
# conda install -c conda-forge "gromacs=2025.4=nompi_h26635d9_100" -y
#
# AmberTools: cpptraj, tleap, parmchk2, sander 제공
# GROMACS (2025.4 nompi): MPI 병렬 실행에 필요 (gmx 명령어)
#   - nompi 버전을 사용해야 mpi4py와 충돌 없음
#   - bioconda의 2021.3 버전은 index 파싱 버그 있음
